refactor: move auth-hook to hooks folder, rename utils
