--- src/navigation/ref.ts
@@ -1,5 +1,0 @@
-// SPDX-License-Identifier: BUSL-1.1
-
-import React from 'react';
-
-export const ref = React.createRef();
