+++ src/utils/application.ts
@@ -1,0 +1,5 @@
+// SPDX-License-Identifier: BUSL-1.1
+
+import {AppState} from 'react-native';
+
+export const isAppActive = () => AppState.currentState === 'active';
